added initial impl of bus directions state

+++ Lazarus/src/com/android/lazarus/serviceadapter/ScheduleServiceAdapter.java
@@ -10,12 +10,12 @@ public interface ScheduleServiceAdapter {
 	 * @return schedule in format hh:mm:ss or null in case of error
 	 */
 	public List<String> getBusSchedule(String token, String lineName,
-			String subLineDescription, int busStopLocationCode,
+			String subLineDescription, long busStopLocationCode,
 			int minutesSinceStartOfDay);
 
 	/**
 	 * @return closest found bus or null
 	 */
-	public Bus getClosestBus(String token, int variantCode, int subLineCode,
-			int busStopOrdinal);
+	public Bus getClosestBus(String token, long variantCode, long subLineCode,
+			long busStopOrdinal);
 }
